// frontend/src/api/generated/models/validationError.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * SpotPass Backend API
 * Restaurant reservation and management system
 * OpenAPI spec version: 1.0.0
 */
import type { ValidationErrorLocItem } from './validationErrorLocItem';

export interface ValidationError {
  loc: ValidationErrorLocItem[];
  msg: string;
  type: string;
}